REPL seems not to spin anymore
The problem seemed to be the following: When closing the graphical repl
the corresponding thread is closed and then cmd-wrt is tagged as
_broken_ (according to Java documentation). We have now (hopefully)
fixed is-eof-ex? to catch the exception thrown ("Write end dead") and
thus the repl is closed on exit. This marks repl-alive? as false and the
output-polling-thread is stopped.

Signed-off-by: Daniel Borchmann <daniel.borchmann@mailbox.tu-dresden.de>