Fixed bug in implementation of explore-attribute
i.e. when a counter-example is added the current pseudo-intent has to be
dealt with again -- changed that. Also added some more tests, also
covering the addition of counter examples.

Note: The semantics of the handler for explore-attributes has been
changed. It is now documented in the doc-string of explore-attributes.

Signed-off-by: Daniel Borchmann <daniel.borchmann@mailbox.tu-dresden.de>