As it seems, the EL_gfp exploration bug is fixed
The reason might be the following: Since classes for DL-expressions and
the like get recompiled every now and then, their "order" in sets
changes accordingly (since this is determined by their hash values,
which are in turn influenced by the underlying class). This causes some
indeterminism in the order of concept names if sets are involved.

Fixed this by explicitly sorting things. This is not the best to do, but
it seems to work, at least for now.

Signed-off-by: Daniel Borchmann <daniel.borchmann@mailbox.tu-dresden.de>