Updating the script
I would like to have c.c.gui.repl-utils/*main-frame* to be
bound to the frame of the gui if it is started.  However,
it seems that the compiler itself does not recognize non-toplevel
require statements and dies.  Thus, I have to require the gui
code regardless of the fact whether the gui is started or not,
which is kind of pitty.

Signed-off-by: Daniel Borchmann <daniel.borchmann@mailbox.tu-dresden.de>