Allow the incidence relation to be a function
Up to now, when given a function as the incidence relation for
make-context, it has been converted to a set.  This is costly and
unnecessary.  Now we allow the incidence relation to stay a function.

This is highly experimental.  All tests work again, but some code is not
covered by tests, and it may fail now.

Signed-off-by: Daniel Borchmann <daniel.borchmann@mailbox.tu-dresden.de>