Fix implementation of concept stability
Depth-first search was done on the wrong structure.  Previously we
considered every element to be in or not, independently of the other
choices; in this way duplication can occur quite easily.  Now we
consider the elements in some (arbitrary) order, ensuring that
duplication cannot happen.  This way, the code is correct and also much
faster.  Yay!

Signed-off-by: Daniel Borchmann <daniel@algebra20.de>