Added concept-probability measure.
Two functions for that are now implemented. The first one,
`concept-probability' computes the exact fraction, however,
is rather slow. The second one utilizes double precision floating
point arithmetic and omits certain factors that are not representable
in this precision anyway. But, in cases where contexts have thounsands
of attributes this function may not able to compute a correct result.