changed context io test to ignore :anonymous-burmeister format
The fix for the :anonymous-burmeister write to file function caused
some context io tests to fail.  This commit removes the
:anonymous-burmeister format from the out-in and out-in-out-in tests.

The reasoning is as follows:

- Before fixing the io write function, the tests went through because
calling `count` on a context results in an
`UnsupportedOperationException` which is caught and accepted (other
formats throw this exception, namely :fcalgs, which only accepts
integral objects and attributes).

- Fixing the write part for the :anonymous-burmeister format meant
that the tests were actually running

- The out-in tests fail for the :anonymous-burmeister format basically
by design because the object and attribute labels are discarded on write

- The out-in-out-in tests usually fail too because the order of
objects and attributes is not guaranteed to be kept intact on
write. This is because `(seq (set (range N)))` is not guaranteed to be
the same as `(range N)`. Hence the context after importing is the same
context but possibly with shuffled objects and attributes, and thus
the equality check fails.