Added possible-isomorphy tests for :anonymous-burmeister format
Previously there were no working tests for the :anonymous-burmeister
format.  Because this format drops object and attribute names when
writing a context to a file and there is no determinism in the order
of objects and attributes written to file, we can not test for
equality in out-in and out-in-out-in tests.  Therefore, we resort to
testing for possible isomorphy -- testing for true isomorphy is not
trivial and expensive to compute.